+++ src/main/java/fr/sylvainjanet/app/App.java
@@ -3,17 +3,9 @@ package fr.sylvainjanet.app;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import fr.sylvainjanet.app.dtos.StringDto;
 
 /**
  * Main App.
@@ -21,7 +13,6 @@ import fr.sylvainjanet.app.dtos.StringDto;
  * @author Sylvain Janet
  *
  */
-@Controller
 @EnableConfigurationProperties
 @SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
 @ConfigurationProperties(prefix = "app")
@@ -34,7 +25,12 @@ public class App extends SpringBootServletInitializer {
   private String environment;
 
   /**
-   * Get the current environment (maven profile).
+   * Get the current environment. Corresponds to the current maven profile.
+   * 
+   * <p>Can be : dev, staging, prod for different stages of development.
+   * Can be coverage-dev, coverage-staging, or coverage-prod during test
+   * coverage computation. Can be depGraph during dependency graph
+   * computation.
    * 
    * @return the environment
    */
@@ -44,25 +40,18 @@ public class App extends SpringBootServletInitializer {
 
   /**
    * Set the current environment. Used for property binding by the
-   * ConfigurationProperties annotation.
+   * ConfigurationProperties annotation. Corresponds to the current maven
+   * profile.
+   * 
+   * <p>Can be : dev, staging, prod for different stages of development.
+   * Can be coverage-dev, coverage-staging, or coverage-prod during test
+   * coverage computation. Can be depGraph during dependency graph
+   * computation.
    * 
    * @param environment the profile
    */
   public void setEnvironment(final String environment) {
     this.environment = environment;
-  }
-
-  /**
-   * Mapping used for basic testing of the app functionality : the API
-   * Hello world.
-   *
-   * @return hello
-   */
-  @GetMapping("/hello")
-  @ResponseBody
-  ResponseEntity<StringDto> home() {
-    return new ResponseEntity<>(
-        new StringDto("Hello World ! - " + environment), HttpStatus.OK);
   }
 
   /**
@@ -73,14 +62,4 @@ public class App extends SpringBootServletInitializer {
   public static void main(final String[] args) {
     SpringApplication.run(App.class, args);
   }
-
-  /**
-   * Build the spring application.
-   */
-  @Override
-  protected SpringApplicationBuilder configure(
-      final SpringApplicationBuilder builder) {
-    return builder.sources(App.class);
-  }
-
 }
