+++ src/main/java/fr/sylvainjanet/app/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Main package.
+ * Main package used to launch and configure the Spring boot application.
  */
 package fr.sylvainjanet.app;
